backend/requirements.txt
django==3.2.21
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
PyJWT==2.8.0
firebase-admin==6.4.0